Move LabyMod netty methods to LabyModChannelUtils
