+++ cloudnet-modules/cloudnet-labymod/src/main/java/eu/cloudnetservice/cloudnet/ext/labymod/LabyModConstants.java
@@ -1,5 +1,10 @@
 package eu.cloudnetservice.cloudnet.ext.labymod;
 
+import de.dytanic.cloudnet.driver.service.ServiceEnvironmentType;
+
+import java.util.Collection;
+import java.util.Collections;
+
 public interface LabyModConstants {
 
     String LMC_CHANNEL_NAME = "LMC";
@@ -7,5 +12,7 @@ public interface LabyModConstants {
     String GET_PLAYER_JOIN_SECRET = "get_player_by_join_secret";
     String GET_PLAYER_SPECTATE_SECRET = "get_player_by_spectate_secret";
     String CLOUDNET_CHANNEL_NAME = "cloudnet_labymod_module";
+
+    Collection<ServiceEnvironmentType> SUPPORTED_ENVIRONMENTS = Collections.singletonList(ServiceEnvironmentType.BUNGEECORD);
 
 }
